@@ -37,29 +37,29 @@
     "quantum-tensors": "^0.4.10",
     "register-service-worker": "^1.6.2",
     "tone": "^13.8.25",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0-rc.10",
     "vue-confetti": "^2.0.5",
     "vue-css-donut-chart": "^1.1.6",
-    "vue-gtag": "^1.6.2",
-    "vue-i18n": "^8.21.0",
-    "vue-router": "^3.0.3",
-    "vuex": "^3.0.1",
+    "vue-i18n": "^9.0.0-alpha.17",
+    "vue-router": "^4.0.0-beta.9",
+    "vuex": "^4.0.0-beta.4",
     "vuex-class": "^0.3.2"
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.0.1",
+    "@typescript-eslint/parser": "^4.0.1",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-typescript": "^4.5.3",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-0",
+    "@vue/eslint-config-typescript": "^5.0.2",
     "@types/jest": "^24.0.18",
     "@types/markdown-it": "^0.0.9",
     "@types/webpack-env": "^1.14.1",
-    "@typescript-eslint/eslint-plugin": "^4.0.1",
-    "@typescript-eslint/parser": "^4.0.1",
-    "@vue/cli-plugin-babel": "^4.1.1",
-    "@vue/cli-plugin-eslint": "^4.1.1",
     "@vue/cli-plugin-pwa": "^4.1.1",
-    "@vue/cli-plugin-typescript": "^4.1.1",
-    "@vue/cli-service": "^4.1.1",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-standard": "^5.0.0",
-    "@vue/eslint-config-typescript": "^5.0.1",
     "babel-eslint": "^10.0.1",
     "eslint": "^6.7.2",
     "eslint-plugin-import": "^2.18.2",
@@ -77,10 +77,7 @@
     "sass-loader": "^8.0.0",
     "ts-node": "^8.8.1",
     "typescript": "^3.4.3",
-    "vue-class-component": "^7.1.0",
-    "vue-property-decorator": "^8.2.2",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.41.2",
-    "webpack-env": "^0.8.0"
+    "vue-class-component": "^8.0.0-alpha.6",
+    "vue-property-decorator": "^8.2.2"
   }
 }
