@@ -1,7 +1,4 @@
-import Vue, { CreateElement, VNode } from 'vue'
-import Donut from 'vue-css-donut-chart'
-import VueConfetti from 'vue-confetti'
-import 'vue-css-donut-chart/dist/vcdonut.css'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store/store'
@@ -9,15 +6,12 @@ import './registerServiceWorker'
 import '@/config/firebase'
 import { i18n } from './i18n'
 
-Vue.use(Donut)
-Vue.use(VueConfetti)
-
-Vue.config.productionTip = false
-new Vue({
-  router,
-  i18n,
-  store,
-  render: (h: CreateElement): VNode => {
-    return h(App)
-  },
-}).$mount('#app')
+createApp(App)
+  .use(router)
+  // FIXME: vue3 support soon: https://github.com/MatteoGabriele/vue-gtag/issues/168#issuecomment-684498923
+  // .use(VueGtag, {
+  //   config: { id: 'UA-56225776-1' },
+  // })
+  .use(i18n)
+  .use(store)
+  .mount('#app')
