@@ -5,13 +5,13 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component, Prop } from 'vue-property-decorator'
+import { Vue, Options } from 'vue-class-component'
+import { Prop } from 'vue-property-decorator'
 
-@Component
 export default class AppButton extends Vue {
   @Prop({ default: 'basic' }) readonly type!: string
   @Prop({ default: true }) readonly inline!: boolean
-  @Prop() readonly overlay!: boolean
+  @Prop({ default: false }) readonly overlay!: boolean
 
   get computeClass(): string[] {
     return [this.type, this.inline ? 'inline' : '', this.overlay ? 'overlay-btn' : '']
