@@ -35,7 +35,7 @@
           :key="`cell-${i}-(${cell.coord.x},${cell.coord.y})-${cell.element.name})`"
           :cell="cell"
           :tileSize="tileSize"
-          @click.native="rotate(cell)"
+          @click="rotate(cell)"
         />
       </svg>
     </div>
@@ -55,7 +55,8 @@
 </template>
 
 <script lang="ts">
-import { Vue, Prop, Component } from 'vue-property-decorator'
+import { Vue, Options } from 'vue-class-component'
+import { Prop } from 'vue-property-decorator'
 import Cell from '@/engine/Cell'
 import Particle from '@/engine/Particle'
 import Grid from '@/engine/Grid'
@@ -69,7 +70,7 @@ import { Vector, Frame, Simulation } from 'quantum-tensors'
 import { PolEnum, DirEnum } from 'quantum-tensors/dist/interfaces'
 import { IStyle } from '@/types'
 
-@Component({
+@Options({
   components: {
     AppPhoton,
     AppCell,
