@@ -17,12 +17,12 @@
 
 <script lang="ts">
 import { Dictionary, groupBy } from 'lodash'
-import { Vue, Component } from 'vue-property-decorator'
+import { Vue, Options } from 'vue-class-component'
 import { ILevel } from '@/engine/interfaces'
 import AppLayout from '@/components/AppLayout.vue'
 import levels from '@/assets/data/levels/index'
 
-@Component({
+@Options({
   components: {
     AppLayout,
   },
