@@ -8,13 +8,13 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component } from 'vue-property-decorator'
+import { Vue, Options } from 'vue-class-component'
 import AppButton from '@/components/AppButton.vue'
 import { namespace } from 'vuex-class'
 
 const user = namespace('userModule')
 
-@Component({
+@Options({
   components: {
     AppButton,
   },
